+++ app/Http/Controllers/Api/V1/EducationLevelController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\V1\StoreEducationLevelRequest;
 use App\Http\Requests\V1\UpdateEducationLevelRequest;
 use App\Http\Resources\V1\EducationLevelResource;
 use App\Models\EducationLevel;
+use App\Models\Organization;
 use App\Repositories\EducationLevel\EducationLevelRepositoryInterface;
 use Illuminate\Http\Response;
 
@@ -37,11 +38,13 @@ class EducationLevelController extends Controller
      * @responseFile responses/education-level/education-levels.json
      *
      * @param SearchEducationLevelRequest $request
+     * @param Organization $suborganization
+     * 
      * @return Response
      */
-    public function index(SearchEducationLevelRequest $request)
+    public function index(SearchEducationLevelRequest $request, Organization $suborganization)
     {
-        return  EducationLevelResource::collection($this->educationLevelRepository->getAll($request->all()));
+        return  EducationLevelResource::collection($this->educationLevelRepository->getAll($suborganization, $request->all()));
     }
 
     /**
@@ -61,9 +64,11 @@ class EducationLevelController extends Controller
      * }
      *
      * @param StoreEducationLevelRequest $request
+     * @param Organization $suborganization
+     * 
      * @return Response
      */
-    public function store(StoreEducationLevelRequest $request)
+    public function store(StoreEducationLevelRequest $request, Organization $suborganization)
     {
         $data = $request->validated();
         $educationLevel = $this->educationLevelRepository->create($data);
@@ -88,11 +93,19 @@ class EducationLevelController extends Controller
      *
      * @responseFile responses/education-level/education-level.json
      *
+     * @param Organization $suborganization
      * @param EducationLevel $educationLevel
+     * 
      * @return Response
      */
-    public function show(EducationLevel $educationLevel)
+    public function show(Organization $suborganization, EducationLevel $educationLevel)
     {
+        if ($suborganization->id !== $educationLevel->organization_id) {
+            return response([
+                'message' => 'Invalid education level or organization',
+            ], 400);
+        }
+
         return response([
             'education-level' => new EducationLevelResource($educationLevel),
         ], 200);
@@ -116,10 +129,12 @@ class EducationLevelController extends Controller
      * }
      *
      * @param UpdateEducationLevelRequest $request
+     * @param Organization $suborganization
      * @param EducationLevel $educationLevel
+     * 
      * @return Response
      */
-    public function update(UpdateEducationLevelRequest $request, EducationLevel $educationLevel)
+    public function update(UpdateEducationLevelRequest $request, Organization $suborganization, EducationLevel $educationLevel)
     {
         $data = $request->validated();
         $isUpdated = $this->educationLevelRepository->update($data, $educationLevel->id);
@@ -152,11 +167,19 @@ class EducationLevelController extends Controller
      *   ]
      * }
      *
+     * @param Organization $suborganization
      * @param EducationLevel $educationLevel
+     * 
      * @return Response
      */
-    public function destroy(EducationLevel $educationLevel)
+    public function destroy(Organization $suborganization, EducationLevel $educationLevel)
     {
+        if ($suborganization->id !== $educationLevel->organization_id) {
+            return response([
+                'message' => 'Invalid education level or organization',
+            ], 400);
+        }
+        
         $isDeleted = $this->educationLevelRepository->delete($educationLevel->id);
 
         if ($isDeleted) {
